Fix for OpenTx 2.1.7 and newer
PRECx formatting was changed in OpenTx 2.1.7 to revert back to how the
formatting happened in OpenTx 2.0 (it had been amended to different
format in 2.1.0). This version will need to co exist with the existing
telem1.lua as users on 2.1.0<2.1.7 will need still to use telem1. This
new file renamed with copter prefix as plan to release a plane version
(which has different flight modes and new wav files).
Have also tidied code, removing a lot of unnecessary commenting and
aligning statements where these were unaligned in telem1.lua.